Move info types
This take the three previously refactored info+flag types
and moves them together.

Before, the flag types had to reach into their parent module to bring
their associated info type into scope, while the info type had to do
the same in the other direction. While this is legal, the tighter
coupling made it natural to place them together.

This has a side effect that all FlagBit constants can be made private
within the new files.

slot/{mod.rs:SlotInfo,flag.rs:CkFlag<SlotInfo>} -> slot/slot_info.rs
slot/{mod.rs:TokenInfo,flag.rs:CkFlag<TokenInfo>} -> slot/token_info.rs
mechanism/{mod.rs:MechanismInfo,flags.rs} -> mechanism/mechanism_info.rs
