+++ cryptoki/src/mechanism/mod.rs
@@ -3,11 +3,10 @@
 //! Data types for mechanisms
 
 pub mod elliptic_curve;
-mod flags;
+mod mechanism_info;
 pub mod rsa;
 
 use crate::error::Error;
-use crate::flag::CkFlags;
 use cryptoki_sys::*;
 use log::error;
 use std::convert::{TryFrom, TryInto};
@@ -16,7 +15,7 @@ use std::fmt::Formatter;
 use std::ops::Deref;
 use std::ptr::null_mut;
 
-use flags::*;
+pub use mechanism_info::MechanismInfo;
 
 #[derive(Copy, Debug, Clone, PartialEq, Eq)]
 // transparent so that a vector of MechanismType should have the same layout than a vector of
@@ -703,187 +702,5 @@ impl TryFrom<psa_crypto::types::algorithm::Algorithm> for Mechanism {
                 Err(Error::NotSupported)
             }
         }
-    }
-}
-
-/// Contains information about a mechanism
-#[derive(Debug, Clone, Copy, Default)]
-pub struct MechanismInfo {
-    min_key_size: usize,
-    max_key_size: usize,
-    flags: CkFlags<Self>,
-}
-
-#[doc(hidden)]
-impl From<CK_MECHANISM_INFO> for MechanismInfo {
-    fn from(val: CK_MECHANISM_INFO) -> Self {
-        Self {
-            min_key_size: val.ulMinKeySize as usize,
-            max_key_size: val.ulMaxKeySize as usize,
-            flags: CkFlags::from(val.flags),
-        }
-    }
-}
-
-impl MechanismInfo {
-    /// The minimum size of the key for the mechanism
-    ///
-    /// **[Conformance](crate#conformance-notes):**
-    /// Whether this is measured in bits or in bytes is mechanism-dependent
-    pub fn min_key_size(&self) -> usize {
-        self.min_key_size
-    }
-
-    /// The maximum size of the key for the mechanism
-    ///
-    /// **[Conformance](crate#conformance-notes):**
-    /// Whether this is measured in bits or in bytes is mechanism-dependent
-    pub fn max_key_size(&self) -> usize {
-        self.max_key_size
-    }
-
-    /// True if the mechanism is performed by the device; false if the mechanism is performed in software
-    pub fn hardware(&self) -> bool {
-        self.flags.contains(HW)
-    }
-
-    /// True if the mechanism can be used to encrypt data
-    ///
-    /// See [`Session::encrypt`](crate::session::Session::encrypt)
-    pub fn encrypt(&self) -> bool {
-        self.flags.contains(ENCRYPT)
-    }
-
-    /// True if the mechanism can be used to decrypt encrypted data
-    ///
-    /// See [`Session::decrypt`](crate::session::Session::decrypt)
-    pub fn decrypt(&self) -> bool {
-        self.flags.contains(DECRYPT)
-    }
-
-    /// True if the mechanism can be used to digest a message
-    ///
-    // TODO See [`Session::digest`](crate::session::Session::digest)
-    pub fn digest(&self) -> bool {
-        self.flags.contains(DIGEST)
-    }
-
-    /// True if the mechanism can be used to digitally sign data
-    ///
-    /// See [`Session::sign`](crate::session::Session::sign)
-    pub fn sign(&self) -> bool {
-        self.flags.contains(SIGN)
-    }
-
-    /// True if the mechanism can be used to digitally data which can be recovered from the signature
-    ///
-    // TODO See [`Session::sign_recover`](crate::session::Session::sign_recover)
-    pub fn sign_recover(&self) -> bool {
-        self.flags.contains(SIGN_RECOVER)
-    }
-
-    /// True if the mechanism can be used to verify a digital signature
-    ///
-    /// See [`Session::verify`](crate::session::Session::verify)
-    pub fn verify(&self) -> bool {
-        self.flags.contains(VERIFY)
-    }
-
-    /// True if the mechanism can be used to [`verify`](crate::session::Session::verify) a digital signature and recover the signed data
-    ///
-    // TODO See [`Session::verify_recover`](crate::session::Session::verify_recover)
-    pub fn verify_recover(&self) -> bool {
-        self.flags.contains(VERIFY_RECOVER)
-    }
-
-    /// True if the mechanism can be used to generate a secret key
-    ///
-    // TODO See [`Session::generate`](crate::session::Session::generate)
-    pub fn generate(&self) -> bool {
-        self.flags.contains(GENERATE)
-    }
-
-    /// True if the mechanism can be used to generate a public/private key pair
-    ///
-    /// See [`Session::generate_key_pair`](crate::session::Session::generate_key_pair))
-    pub fn generate_key_pair(&self) -> bool {
-        self.flags.contains(GENERATE_KEY_PAIR)
-    }
-
-    /// True if the mechanism can be used to wrap (encrypt) a key
-    ///
-    // TODO See [`Session::wrap`](crate::session::Session::wrap))
-    pub fn wrap(&self) -> bool {
-        self.flags.contains(WRAP)
-    }
-
-    /// True if the mechanism can be used to unwrap (decrypt) a key
-    ///
-    // TODO See [`Session::unwrap`](crate::session::Session::unwrap))
-    pub fn unwrap(&self) -> bool {
-        self.flags.contains(UNWRAP)
-    }
-
-    /// True if the mechanism can be used to derive a key from a base key
-    ///
-    // TODO See [`Session::derive`](crate::session::Session::derive))
-    pub fn derive(&self) -> bool {
-        self.flags.contains(DERIVE)
-    }
-
-    /// True if there is an extension to the flags; false if no extensions
-    ///
-    /// **[Conformance](crate#conformance-notes):**
-    /// This *must* be false for PKCS#11 v2.40
-    pub fn extension(&self) -> bool {
-        self.flags.contains(EXTENSION)
-    }
-
-    /// True if the mechanism can be used to  with elliptic curve domain parameters over ***F<sub>p</sub>***
-    ///
-    /// **[Conformance](crate#conformance-notes):**
-    /// *At least* one of [`ec_f_p`](Self::ec_f_p) and [`ec_f_2m`](Self::ec_f_2m) must be `true`
-    pub fn ec_f_p(&self) -> bool {
-        self.flags.contains(EC_F_P)
-    }
-
-    /// True if the mechanism can be used with elliptic curve domain parameters over ***F<sub>2<sup>m</sup></sub>***
-    ///
-    /// **[Conformance](crate#conformance-notes):**
-    /// *At least* one of [`ec_f_p`](Self::ec_f_p) and [`ec_f_2m`](Self::ec_f_2m) must be `true`
-    pub fn ec_f_2m(&self) -> bool {
-        self.flags.contains(EC_F_2M)
-    }
-
-    /// True if the mechanism supports specifying elliptic curve domain parameters explicitly
-    ///
-    /// **[Conformance](crate#conformance-notes):**
-    /// *At least* one of [`ec_from_parameters`](Self::ec_from_parameters) and [`ec_from_named_curve`](Self::ec_from_named_curve) must be `true`
-    pub fn ec_from_parameters(&self) -> bool {
-        self.flags.contains(EC_ECPARAMETERS)
-    }
-
-    /// True if the mechanism supports specifying elliptic curve domain parameters with a named curve
-    ///
-    /// **[Conformance](crate#conformance-notes):**
-    /// *At least* one of [`ec_from_parameters`](Self::ec_from_parameters) and [`ec_from_named_curve`](Self::ec_from_named_curve) must be `true`
-    pub fn ec_from_named_curve(&self) -> bool {
-        self.flags.contains(EC_NAMEDCURVE)
-    }
-
-    /// True if the mechanism can be used with elliptic curve points in uncompressed form
-    ///
-    /// **[Conformance](crate#conformance-notes):**
-    /// *At least* one of [`ec_uncompressed`](Self::ec_uncompressed) and [`ec_compressed`](Self::ec_compressed) must be `true`
-    pub fn ec_uncompressed(&self) -> bool {
-        self.flags.contains(EC_UNCOMPRESS)
-    }
-
-    /// True if the mechanism can be used with elliptic curve points in compressed form
-    ///
-    /// **[Conformance](crate#conformance-notes):**
-    /// *At least* one of [`ec_uncompressed`](Self::ec_uncompressed) and [`ec_compressed`](Self::ec_compressed) must be `true`
-    pub fn ec_compressed(&self) -> bool {
-        self.flags.contains(EC_COMPRESS)
     }
 }
